backend/agent/requirements.txt
google-adk
fastapi
uvicorn
requests
pydantic
python-dotenv
toolbox-core
